+++ src/main/kotlin/mc/tsukimiya/mooney/core/domain/Money.kt
@@ -7,11 +7,11 @@ import mc.tsukimiya.mooney.core.exception.InvalidMoneyAmountException
  *
  * @property amount
  */
-data class Money(val amount: Int) {
+data class Money(val amount: ULong) {
     init {
         // amountが0未満なら例外投げる
         // システムに0未満の値を存在させないため
-        if (amount < 0) {
+        if (amount < 0u) {
             throw InvalidMoneyAmountException("Wrong amount : $amount | Require $amount >= 0")
         }
     }
