+++ src/main/kotlin/mc/tsukimiya/mooney/core/usecase/dto/AccountDto.kt
@@ -2,4 +2,4 @@ package mc.tsukimiya.mooney.core.usecase.dto
 
 import java.util.*
 
-data class AccountDto(val id: UUID, val name: String, val money: Int)
+data class AccountDto(val id: UUID, val name: String, val money: ULong)
